+++ docs/nontransparent.dot
@@ -1,4 +1,4 @@
-digraph rfc6587 {
+digraph nontransparent {
 	rankdir=LR;
 	node [ shape = point ];
 	ENTRY;
@@ -8,11 +8,11 @@ digraph rfc6587 {
 	3;
 	node [ shape = circle ];
 	1 -> 2 [ label = "'<' / on_init" ];
-	2 -> 3 [ label = "'\\n' / on_trailer" ];
-	2 -> 2 [ label = "DEF" ];
-	3 -> 3 [ label = "'\\n' / on_trailer" ];
+	2 -> 2 [ label = "1..'\\t', '\\v'..255, '\\n'(!27:13), 0(!27:13)" ];
+	2 -> 3 [ label = "'\\n'(27:13), 0(28:13) / on_trailer" ];
+	3 -> 2 [ label = "1..'\\t', '\\v'..';', '='..255, '\\n'(!27:13), 0(!27:13)" ];
 	3 -> 2 [ label = "'<' / on_init" ];
-	3 -> 2 [ label = "DEF" ];
+	3 -> 3 [ label = "'\\n'(27:13), 0(28:13) / on_trailer" ];
 	ENTRY -> 1 [ label = "IN" ];
 	en_1 -> 1 [ label = "main" ];
 }
